bgs: New function tick_time_x_ago()
This determines the time of the tick X ticks ago.  Being used in place
of a naive "give me the Xth tick going back in time".

elitebgs.app had some error with tick detection and deleted the last two
days worth.  So then asking for 7 days of ticks and wanting the 6th back
in time was an error, as there were only 5 ticks returned.